% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_convert_pave_to_genic}
\alias{mutation_types_convert_pave_to_genic}
\title{Convert PAVE Mutation Types to Genic}
\usage{
mutation_types_convert_pave_to_genic(
  pave_mutation_types,
  split_on_ampersand = TRUE,
  missing_to_intergenic = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pave_mutation_types}{a vector of PAVE terms you want to convert to genic/intergenic classifications (character)}

\item{split_on_ampersand}{should '&' separated PAVE terms be automatically converted to single PAVE terms based on highest severity? (flag)}

\item{verbose}{verbose (flag)}
}
\value{
whether each variant is 'genic' or 'intergenic' (character)
}
\description{
Convert PAVE mutation types to whether a variant is 'genic' or 'intergenic'
}
\examples{
mutation_types_convert_pave_to_maf(
  c('upstream_gene_variant', 'stop_lost', 'splice_acceptor_variant')
)
}
