% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_identify}
\alias{mutation_types_identify}
\title{Identify Mutation Dictionary Used}
\usage{
mutation_types_identify(mutation_types, verbose = TRUE)
}
\arguments{
\item{mutation_types}{mutation types to test (character)}

\item{verbose}{verbose (flag)}
}
\value{
one of c('SO', 'MAF', 'UNKNOWN').
Will return 'UNKNOWN' unless ALL mutation types fit with one of the supported dictionaries
}
\description{
Identify Mutation Dictionary Used
}
\examples{
mutation_types_identify(c('bob', 'billy', 'missense_variant'))
}
