% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_maf}
\alias{mutation_types_maf}
\title{Dictionary of MAF terms}
\usage{
mutation_types_maf()
}
\value{
valid MAF terms (character)
}
\description{
Dictionary of MAF terms
}
\examples{
mutation_types_maf()
}
