% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_maf_palette}
\alias{mutation_types_maf_palette}
\title{Palettes: MAF}
\usage{
mutation_types_maf_palette()
}
\value{
named vector. Names are MAF terms. Values are colors
}
\description{
Palettes: MAF
}
\examples{
mutation_types_maf_palette()
}
