% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_pave}
\alias{mutation_types_pave}
\title{Dictionary of PAVE terms}
\usage{
mutation_types_pave()
}
\value{
valid PAVE terms (character)
}
\description{
PAVE is a newer annotation software that supports annotaiton of mainly just a subset of SO terms,
but with a couple of important additions to indicate when a non-obvious consequence can be found thanks to phasing.
}
\examples{
mutation_types_pave()
}
