% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_pave_palette}
\alias{mutation_types_pave_palette}
\title{Palettes: PAVE}
\usage{
mutation_types_pave_palette()
}
\value{
named vector. Names are PAVE terms. Values are colors
}
\description{
Palettes: PAVE
}
\examples{
mutation_types_pave_palette()
}
