% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_so}
\alias{mutation_types_so}
\title{Dictionary of So terms}
\usage{
mutation_types_so()
}
\value{
valid SO terms (character)
}
\description{
Dictionary of So terms
}
\examples{
mutation_types_so()
}
