% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_so_palette}
\alias{mutation_types_so_palette}
\title{Palettes: SO}
\usage{
mutation_types_so_palette()
}
\value{
named vector. Names are SO terms. Values are colors
}
\description{
Palettes: SO
}
\examples{
mutation_types_so_palette()
}
