% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationtypes-package.R
\docType{package}
\name{mutationtypes-package}
\alias{mutationtypes}
\alias{mutationtypes-package}
\title{mutationtypes: Ensure Mutational Impacts Conform to Standard Dictionaries}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Check concordance of a vector of mutation impacts with 'SO' or 'MAF' dictionaries. Can also convert SO terms to MAF terms and/or select the most severe SO consequence from ampersand (&) delimited impacts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/mutationtypes}
  \item Report bugs at \url{https://github.com/selkamand/mutationtypes/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{sam.elkamand@gmail.com} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID}) [copyright holder]

}
\keyword{internal}
