% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{select_most_severe_consequence_so}
\alias{select_most_severe_consequence_so}
\title{Select the most severe consequence (SO)}
\usage{
select_most_severe_consequence_so(so_mutation_types_list)
}
\arguments{
\item{so_mutation_types_list}{a list, where each element is a vector containing the set of SO terms you want to choose the most severe consequence from.}
}
\value{
the most severe consequence for each vector in so_mutation_types_list
}
\description{
Select the most severe consequence (SO)
}
\examples{
select_most_severe_consequence_so(
  list(
     c("intergenic_variant", "feature_truncation", "splice_acceptor_variant"),
     c("initiator_codon_variant", "inframe_insertion")
  )
)
#> Result:
#> c("splice_acceptor_variant", "initiator_codon_variant")
}
