% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{mutation_types_convert_so_to_maf}
\alias{mutation_types_convert_so_to_maf}
\title{Convert SO Mutation Types to MAF}
\usage{
mutation_types_convert_so_to_maf(so_mutation_types)
}
\arguments{
\item{so_mutation_types}{a vector of SO terms you want to convert to MAF variant classifications}
}
\value{
(character)
}
\description{
Convert SO Mutation Types to MAF
}
\examples{
mutation_types_convert_so_to_maf(c('INTRAGENIC', 'INTRAGENIC', 'intergenic_region'))
}
