% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationtypes-package.R
\docType{package}
\name{mutationtypes-package}
\alias{mutationtypes}
\alias{mutationtypes-package}
\title{mutationtypes: Validate and Convert Mutational Impacts Using Standard Genomic Dictionaries}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Check concordance of a vector of mutation impacts with standard dictionaries such as Sequence Ontology (SO) \url{http://www.sequenceontology.org/}, Mutation Annotation Format (MAF) \url{https://docs.gdc.cancer.gov/Encyclopedia/pages/Mutation_Annotation_Format_TCGAv2/} or Prediction and Annotation of Variant Effects (PAVE) \url{https://github.com/hartwigmedical/hmftools/tree/master/pave}. It enables conversion between SO/PAVE and MAF terms and selection of the most severe consequence where multiple ampersand (&) delimited impacts are given.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/mutationtypes}
  \item Report bugs at \url{https://github.com/selkamand/mutationtypes/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{sam.elkamand@gmail.com} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

Other contributors:
\itemize{
  \item Children's Cancer Institute Australia [copyright holder]
}

}
\keyword{internal}
